+++ convert/LongConverter.go
@@ -38,8 +38,31 @@ func (c *TLongConverter) ToLong(value interface{}) int64 {
 	return ToLong(value)
 }
 
+// Converts value into unsigned long or returns default when conversion is not possible.
+// Parameters:
+//  "value" - the value to convert.
+//  "defaultValue" - the default value..
+// Returns: long value or default when conversion is not supported.
+func (c *TLongConverter) ToULongWithDefault(value interface{}, defaultValue uint64) uint64 {
+	return ToULongWithDefault(value, defaultValue)
+}
+
+// Converts value into unsigned long or returns null when conversion is not possible.
+// Parameters: "value" - the value to convert
+// Returns: long value or null when conversion is not supported.
+func (c *TLongConverter) ToNullableULong(value interface{}) *uint64 {
+	return ToNullableULong(value)
+}
+
+// Converts value into unsigned long or returns 0 when conversion is not possible.
+// Parameters: "value" - the value to convert
+// Returns: long value or 0 when conversion is not supported.
+func (c *TLongConverter) ToULong(value interface{}) uint64 {
+	return ToULong(value)
+}
+
 // Converts value into long or returns default when conversion is not possible.
-// Parameters: 
+// Parameters:
 //  "value" - the value to convert.
 //  "defaultValue" - the default value..
 // Returns: long value or default when conversion is not supported.
@@ -115,12 +138,92 @@ func ToLong(value interface{}) int64 {
 }
 
 // Converts value into long or returns default when conversion is not possible.
-// Parameters: 
+// Parameters:
 //  "value" - the value to convert.
 //  "defaultValue" - the default value..
 // Returns: long value or default when conversion is not supported.
 func ToLongWithDefault(value interface{}, defaultValue int64) int64 {
 	r := ToNullableLong(value)
+	if r == nil {
+		return defaultValue
+	}
+	return *r
+}
+
+// Converts value into unsigned long or returns null when conversion is not possible.
+// Parameters: "value" - the value to convert
+// Returns: long value or null when conversion is not supported.
+func ToNullableULong(value interface{}) *uint64 {
+	if value == nil {
+		return nil
+	}
+
+	var r uint64 = 0
+
+	switch value.(type) {
+	case int8:
+		r = (uint64)(value.(int8))
+	case uint8:
+		r = (uint64)(value.(uint8))
+	case int:
+		r = (uint64)(value.(int))
+	case int16:
+		r = (uint64)(value.(int16))
+	case uint16:
+		r = (uint64)(value.(uint16))
+	case int32:
+		r = (uint64)(value.(int32))
+	case uint32:
+		r = (uint64)(value.(uint32))
+	case int64:
+		r = (uint64)(value.(int64))
+	case uint64:
+		r = (uint64)(value.(uint64))
+	case float32:
+		r = (uint64)(value.(float32))
+	case float64:
+		r = (uint64)(value.(float64))
+
+	case bool:
+		v := value.(bool)
+		if v == true {
+			r = 1
+		}
+
+	case time.Time:
+		r = (uint64)(value.(time.Time).Unix())
+
+	case time.Duration:
+		r = (uint64)(value.(time.Duration).Nanoseconds() / 1000000)
+
+	case string:
+		v, ok := strconv.ParseUint(value.(string), 10, 0)
+		if ok != nil {
+			return nil
+		}
+		r = uint64(v)
+
+	default:
+		return nil
+	}
+
+	return &r
+}
+
+// Converts value into unsigned long or returns 0 when conversion is not possible.
+// Parameters: "value" - the value to convert
+// Returns: long value or 0 when conversion is not supported.
+func ToULong(value interface{}) uint64 {
+	return ToULongWithDefault(value, 0)
+}
+
+// Converts value into unsigned long or returns default when conversion is not possible.
+// Parameters:
+//  "value" - the value to convert.
+//  "defaultValue" - the default value..
+// Returns: long value or default when conversion is not supported.
+func ToULongWithDefault(value interface{}, defaultValue uint64) uint64 {
+	r := ToNullableULong(value)
 	if r == nil {
 		return defaultValue
 	}