+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Portable Abstractions and Patterns for Golang Changelog
 
+## <a name="1.1.1"></a> 1.1.1 (2021-05-12)
+
+### Bug Fixing
+
+* Fix converting from uint64 to string and from string to uint64
+
+### Features
+
+* Added separated methods for conversion into signed and unsigned values 
+
 ## <a name="1.1.0"></a> 1.1.0 (2021-04-03)
 
 ### Features